Placeholder config file for key metrics
All components are currently commented out of compute_notebooks section, but as
we create notebooks we can add them here. Note that we want to provide plots
highlighting global run .092, so that case is specified as a global parameter.
If we want notebooks to compare to other cases, we will need a way to specify
those other output directories as well.