Use base_case_output_dir in timeseries
If one of the case names in the timeseries section of the config file matches
base_case_name, look for output in base_case_output_dir instead of
CESM_output_dir.

Also, one of the strings used when raising an exception was meant to be an
f-string (it includes a {variable} but was missing the f").