Add support for CVDP in gen_adf_config
generate_adf_config_file.py accepts a --run-cvdp option, which will set the
cvdp_run to True in adf_config.yml. This might not be the best implementation,
I think the CUPiD config.yml will still show "cvdp_run: false". It may make
more sense to have this flag in generate_cupid_config; I'll play with that in a
subsequent commit.

Also added CUPID_RUN_CVDP to config_tool.xml; I think we need a check to make
sure we don't run with CUPID_RUN_CVDP=TRUE and CUPID_RUN_ADF=FALSE, since that
would result in CVDP not being run (it is launched as part of ADF).

Lastly, cleaned up some comments in the description of variables of
config_tool.xml -- CUPID_RUN_ADF and its ilk are independent of CUPID_RUN_ALL
(the latter just applies to which components we run diagnostics on).